# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GamiBook API",
        "description": "ESMAPP API - GamiBook é uma aplicação para a editora LIDEL que incentiva a aprendizagem através de exercícios gamificados. O token não expira já que a aplicação é destinada para o uso mobile.",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
        "parameters": {
            "idNotification": {
                "name": "idNotification",
                "in": "path",
                "description": "Identificador da notificação",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "idUser": {
                "name": "idUser",
                "in": "path",
                "description": "Identificador do utilizador",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
...
        },
...
    },
...
    "paths": {
...
        "/users": {
            "get": {
                "summary": "Lista de utilizadores",
                "description": "Devolve uma lista de utilizadores da aplicação",
                "operationId": "findAllUsers",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/findAllUsers"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            },
            "post": {
                "summary": "Criar utilizadores administradores",
                "description": "Adiciona um utilizador à base de dados da aplicação",
                "operationId": "createAdminUser",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/users/register_admin_user"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/users/201/createAdminUser"
                    },
                    "400": {
                        "$ref": "#/components/responses/users/400/createAdminUser"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "summary": "Perfil do utilizador",
                "description": "Devolve a informação do perfil do utilizador",
                "operationId": "getProfile",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Geral"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/getProfile"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    }
                }
            },
            "patch": {
                "summary": "Atualizar o perfil",
                "description": "Atualiza certos campos da informação do utilizador",
                "operationId": "updateProfile",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Geral"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/users/updateProfile"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/updateProfile"
                    },
                    "400": {
                        "$ref": "#/components/responses/users/400/updateProfile"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    }
                }
            }
        },
        "/users/me/notifications": {
            "get": {
                "summary": "Notificações do utilizador",
                "description": "Devolve uma lista de notificações do utilizador",
                "operationId": "findProfileNotifications",
                "tags": [
                    "Notificação",
                    "Rotas Privadas - Utilizador Regular"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/notifications/200/findProfileNotifications"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            }
        },
        "/users/me/notifications/{idNotification}": {
            "delete": {
                "summary": "Apagar notificações",
                "description": "Apaga uma notificação pelo id",
                "operationId": "deleteProfileNotification",
                "tags": [
                    "Notificação",
                    "Rotas Privadas - Utilizador Regular"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idNotification"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/notifications/200/deleteProfileNotification"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idNotification"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idNotification"
                    }
                }
            }
        },
        "/users/{idUser}": {
            "get": {
                "summary": "Procurar um utilizador",
                "description": "Devolve a informação de um utilizador pelo id",
                "operationId": "findOneUser",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idUser"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/findOneUser"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idUser"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idUser"
                    }
                }
            }
        },
...
    }
}